screenshot-capture-tool/requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
mss>=9.0.0
PyInstaller>=6.0.0
